fix: test with schedule

+++ index.js
@@ -6,8 +6,6 @@ const FormData = require('form-data')
 
 const client = new discord.Client()
 
-console.log('Handle ran.', process.env.DISCORD_TOKEN)
-
 client.login(process.env.DISCORD_TOKEN)
 
 client.once('ready', async () => {
@@ -31,11 +29,11 @@ client.once('ready', async () => {
       if (response.status === 200) {
         const hasInitalForm = response.data.includes('action="/booking/create/48803/0"')
         const valid = !response.data.includes('Il n\'existe plus de plage horaire libre pour votre demande')
-        console.log('Validity', hasInitalForm, valid)
-        // if (!hasInitalForm && valid) {
-        await channel.send('Un créneau semble disponible ! https://www.seine-et-marne.gouv.fr/booking/create/48803/2')
-        await client.destroy()
-        // }
+        if (!hasInitalForm && valid) {
+          console.log('Available!')
+          await channel.send('Un créneau semble disponible ! https://www.seine-et-marne.gouv.fr/booking/create/48803/2')
+          await client.destroy()
+        }
         process.exit(0)
       }
     })
